Add emptyToNull and blankToNull functions
Motivation:

When a mapping converts a JSON String to a predicate-object pair,
RMLMapper will decline to generate a corresponding triple for JSON
Strings that are `null`.  An empty JSON String value (`""`) is
considered valid and is mapped.

In some cases, an empty string actually indicates that the value is
missing.  This may be because `null` is consciously avoided or from poor
data validation.

Modification:

Add two new methods for handling non-null data that should not generate
triples: `emptyToNull` and `blankToNull`.

Results:

Handle string fields in which the absense of data is represented by an
empty string or (more generally) as a string containing only whitespace.