Add support for leading zeros
Motivation:

Some formats require data to have a specific size.  For example,
xsd:gYear is defined to be an integer value YYYY (exactly four digits);
therefore, the year 971 must be written 0971.

Modification:

Add new function 'withLeadingZeros' that builds a String representation
of an integer number, but with a specified minimum number of digits.
Input with too few digits are zero-padded.

Result:

It is now possible to satisfy formats with fixed-width requirements.