Add missing dcterms prefix in functions definition.
Motivation:

The function definitions supplied with rml-extra-functions uses Dublin
Core terms when describing functions.  This is (mistakenly) done without
defining the `dcterms` prefix.

This works-by-accident with rmlmapper v5.0.0 and earlier, but fails with
v6.0.0 and newer.

Modification:

The missing prefix definition for the 'dcterms' is added to the function
definitions.

Result:

Newer rmlmapper (v6.0.0 and newer) no longer complains about the missing
dcterms prefix.  This change does not affect earlier versions of
rmlmapper.