Switch `Mapping` namespace from `fnoi` to `fno`
Motivation:

rmlmapper commit a52f1d7168 updated the function-agent-java library from
v0.0.1 to v0.0.2.  This meant that, since v6.0.0, rmlmapper now uses the
newer version of this dependency.

The new version of function-agent-java included a fix for the `Mapping`
term.  In v0.0.1, the library processed the `Mapping` term from the
`fnoi` namespace.  This is, however, wrong: the correct namespace is the
implementation-agnostic `fno` namespace.

The function definitions for rml-extra-functions currently uses the
incorrect fnoi namespace.  This prevents the functions from being used
with rmlmapper v6.0.0 or newer.

Modification:

Update function declarations to use the correct namespace for the
`Mapping` terms.

Result:

The function definitions are now compatible with rmlmapper v6.0.0 (or
newer).

By changing the namespace of the `Mapping` terms, the `functions.ttl`
file is now incompatible with earlier versions of rmlmapper, those
before v6.0.0.  Although it might be possible to write a `functions.ttl`
file that is backwards-compatible with rmlmapper prior to v6.0.0, I
think it would be easier simply simply to require a newer version of
rmlmapper.

My thanks go to @pulquero for pointing out this problem in #1.

Fixes #1